requirements.txt
fastapi==0.115.12
markdown_pdf==1.3.3
openai==1.69.0
pandas==2.2.0
pydantic==2.11.0
pypdf==5.2.0
pypdfium2==4.0.0
python-dotenv==1.0.1
python_docx==1.1.2
requests==2.32.3
striprtf==0.0.28
fastapi-cli==0.0.7
python-multipart==0.0.20
xlrd==2.0.1
openpyxl==3.1.0
boto3==1.37.23
uuid==1.30
librosa
soundfile
numpy
scipy